Rename ipywidgets_bokeh/{ipy_->}widget.py
