+++ ipywidgets_bokeh/__init__.py
@@ -1,3 +1,3 @@
-from .ipy_widget import IPyWidget
+from .widget import IPyWidget
 
 __version__ = "1.0.0dev1"
